add ili9225 display

+++ lib/print/printf.cpp
@@ -79,7 +79,7 @@ void PrintF::printf(const char *string, ...)
       case 'u':
         switch (lng) {
         #ifdef NUM_16
-          case 2: ptr = sprint(ptr, (uint16_t)__builtin_va_arg(args, void *)); break;
+          case 2: ptr = sprint(ptr, (uint16_t)(uint32_t)__builtin_va_arg(args, void *)); break;
           #endif
           #ifdef NUM_32
           case 4: ptr = sprint(ptr, (uint32_t)__builtin_va_arg(args, uint32_t)); break;
@@ -94,7 +94,7 @@ void PrintF::printf(const char *string, ...)
       case 'd':
         switch (lng) {
         #ifdef NUM_16
-          case 2: ptr = sprint(ptr, (int16_t)__builtin_va_arg(args, void *)); break;
+          case 2: ptr = sprint(ptr, (int16_t)(uint32_t)__builtin_va_arg(args, void *)); break;
           #endif
           #ifdef NUM_32
           case 4: ptr = sprint(ptr, (int32_t)__builtin_va_arg(args, int32_t)); break;
@@ -109,7 +109,7 @@ void PrintF::printf(const char *string, ...)
       case 'q':
         switch (lng) {
         #ifdef NUM_16
-          case 2: ptr = sprint(ptr, (int16_t)__builtin_va_arg(args, void *), fix, fix2); break;
+          case 2: ptr = sprint(ptr, (int16_t)(uint32_t)__builtin_va_arg(args, void *), fix, fix2); break;
           #endif
           #ifdef NUM_32
           case 4: ptr = sprint(ptr, (int32_t)__builtin_va_arg(args, int32_t), fix, fix2); break;
